+++ crates/jmap/src/registry/mapping/task.rs
@@ -399,7 +399,7 @@ pub(crate) async fn task_get(
 pub(crate) async fn task_query(
     mut req: RegistryQueryResponse<'_>,
 ) -> trc::Result<QueryResponseBuilder> {
-    let mut due_from = 100u64;
+    let mut due_from = 1u64;
     let mut due_to = u64::MAX;
     let mut typ = None;
 
@@ -466,6 +466,7 @@ pub(crate) async fn task_query(
         .extract_parameters(req.server.core.jmap.query_max_results, None)?;
 
     let mut from_id = 0u64;
+    let mut to_id = u64::MAX;
     if let Some(anchor_id) = anchor_id
         && let Some(anchor_task) = req
             .server
@@ -478,8 +479,13 @@ pub(crate) async fn task_query(
     {
         let anchor_due = anchor_task.due_timestamp();
         if anchor_due >= due_from && anchor_due <= due_to {
-            due_from = anchor_due;
-            from_id = anchor_id;
+            if params.sort_ascending {
+                due_from = anchor_due;
+                from_id = anchor_id;
+            } else {
+                due_to = anchor_due;
+                to_id = anchor_id;
+            }
         }
     }
 
@@ -497,7 +503,7 @@ pub(crate) async fn task_query(
         due: due_from,
     }));
     let to_key = ValueKey::from(ValueClass::TaskQueue(TaskQueueClass::Due {
-        id: u64::MAX,
+        id: to_id,
         due: due_to,
     }));